got rid of all the junk and organized folders/data
