+++ src/basic/etlBasic.js
@@ -2,27 +2,28 @@ const fs = require('fs');
 const path = require('path');
 const csv = require('fast-csv');
 const mongoose = require('mongoose');
-const { addReviews } = require('./etlBasicHelpers.js');
+const { addReviews, addPhotos } = require('./etlBasicHelpers.js');
 
-const reviewsCSV = './basicData/reviews100K.csv';
+const reviewsCSV = '../../data-50k/reviews50k.csv';
+const photosCSV = '../../data-50k/photos50k.csv';
 
-let operations = [];
-
+// Need to add a port number?
 mongoose.connect('mongodb://127.0.0.1/reviews',
   {
     useNewUrlParser: true,
     useUnifiedTopology: true,
-    // useCreateIndex: true,
-    // autoIndex: true
   })
   .then(() => {
     console.log(`MongoDB Connected!`);
   })
   .then(() => {
     console.log(`Adding Reviews`);
-    addReviews(reviewsCSV);
+    // addReviews(reviewsCSV);
+  })
+  .then(() => {
+    console.log(`Adding Photos`);
+    addPhotos(photosCSV);
   })
   .catch((err) => {
     console.log(`MongoDB ERR ${err}`);
   });
-
