--- src/data/testData/charsTest.csv
@@ -1,186 +1,0 @@
-id,product_id,name
-1,1,Fit
-2,1,Length
-3,1,Comfort
-4,1,Quality
-5,2,Quality
-6,3,Fit
-7,3,Length
-8,3,Comfort
-9,3,Quality
-10,4,Fit
-11,4,Length
-12,4,Comfort
-13,4,Quality
-14,5,Size
-15,5,Width
-16,5,Comfort
-17,5,Quality
-18,6,Size
-19,6,Width
-20,6,Comfort
-21,6,Quality
-22,7,Size
-23,7,Width
-24,7,Comfort
-25,7,Quality
-26,8,Size
-27,8,Width
-28,8,Comfort
-29,8,Quality
-30,9,Size
-31,9,Width
-32,9,Comfort
-33,9,Quality
-34,10,Quality
-35,11,Size
-36,11,Width
-37,11,Comfort
-38,11,Quality
-39,12,Fit
-40,12,Length
-41,12,Comfort
-42,12,Quality
-43,13,Fit
-44,13,Length
-45,13,Comfort
-46,13,Quality
-47,14,Fit
-48,14,Length
-49,14,Comfort
-50,14,Quality
-51,15,Fit
-52,15,Length
-53,15,Comfort
-54,15,Quality
-55,16,Fit
-56,16,Length
-57,16,Comfort
-58,16,Quality
-59,17,Fit
-60,17,Length
-61,17,Comfort
-62,17,Quality
-63,18,Fit
-64,18,Length
-65,18,Comfort
-66,18,Quality
-67,19,Fit
-68,19,Length
-69,19,Comfort
-70,19,Quality
-71,20,Fit
-72,20,Length
-73,20,Comfort
-74,20,Quality
-75,21,Size
-76,21,Width
-77,21,Comfort
-78,21,Quality
-79,22,Quality
-80,23,Size
-81,23,Width
-82,23,Comfort
-83,23,Quality
-84,24,Fit
-85,24,Length
-86,24,Comfort
-87,24,Quality
-88,25,Size
-89,25,Width
-90,25,Comfort
-91,25,Quality
-92,26,Fit
-93,26,Length
-94,26,Comfort
-95,26,Quality
-96,27,Fit
-97,27,Length
-98,27,Comfort
-99,27,Quality
-100,28,Size
-101,28,Width
-102,28,Comfort
-103,28,Quality
-104,29,Fit
-105,29,Length
-106,29,Comfort
-107,29,Quality
-108,30,Fit
-109,30,Length
-110,30,Comfort
-111,30,Quality
-112,31,Fit
-113,31,Length
-114,31,Comfort
-115,31,Quality
-116,32,Fit
-117,32,Length
-118,32,Comfort
-119,32,Quality
-120,33,Quality
-121,34,Fit
-122,34,Length
-123,34,Comfort
-124,34,Quality
-125,35,Fit
-126,35,Length
-127,35,Comfort
-128,35,Quality
-129,36,Quality
-130,37,Fit
-131,37,Length
-132,37,Comfort
-133,37,Quality
-134,38,Fit
-135,38,Length
-136,38,Comfort
-137,38,Quality
-138,39,Fit
-139,39,Length
-140,39,Comfort
-141,39,Quality
-142,40,Fit
-143,40,Length
-144,40,Comfort
-145,40,Quality
-146,41,Fit
-147,41,Length
-148,41,Comfort
-149,41,Quality
-150,42,Fit
-151,42,Length
-152,42,Comfort
-153,42,Quality
-154,43,Size
-155,43,Width
-156,43,Comfort
-157,43,Quality
-158,44,Fit
-159,44,Length
-160,44,Comfort
-161,44,Quality
-162,45,Fit
-163,45,Length
-164,45,Comfort
-165,45,Quality
-166,46,Fit
-167,46,Length
-168,46,Comfort
-169,46,Quality
-170,47,Fit
-171,47,Length
-172,47,Comfort
-173,47,Quality
-174,48,Fit
-175,48,Length
-176,48,Comfort
-177,48,Quality
-178,49,Fit
-179,49,Length
-180,49,Comfort
-181,49,Quality
-182,50,Fit
-183,50,Length
-184,50,Comfort
-185,50,Quality
