--- src/data/testData/exampleData.js
@@ -1,30 +1,0 @@
-[
-  {
-    _id: ObjectId("62a36d9361f6abf17d2a30c9"),
-    product_id: '1',
-    __v: 0,
-    meta: {
-      recommends: { '0': 1, '1': 1 },
-      characteristics: {
-        '1': { name: 'Fit', value: [ 4, 4 ] },
-        '2': { name: 'Length', value: [ 3, 4 ] },
-        '3': { name: 'Comfort', value: [ 5, 5 ] },
-        '4': { name: 'Quality', value: [ 4, 4 ] }
-      }
-    },
-    results: [
-      {
-        id: '1',
-        rating: 5,
-        _id: ObjectId("62a36d932856637fffeb6360"),
-        photos: []
-      },
-      {
-        id: '2',
-        rating: 4,
-        _id: ObjectId("62a36d932856637fffeb6362"),
-        photos: []
-      }
-    ]
-  }
-]
