--- src/data/testData/reviewCharTest.csv
@@ -1,110 +1,0 @@
-id,characteristic_id,review_id,value
-1,1,1,4
-2,2,1,3
-3,3,1,5
-4,4,1,4
-5,1,2,4
-6,2,2,4
-7,3,2,5
-8,4,2,4
-9,5,3,4
-10,5,4,5
-11,5,5,5
-12,5,6,3
-13,5,7,4
-14,10,8,2
-15,11,8,3
-16,12,8,2
-17,13,8,1
-18,10,9,4
-19,11,9,3
-20,12,9,4
-21,13,9,5
-22,10,10,5
-23,11,10,5
-24,12,10,5
-25,13,10,5
-26,14,11,5
-27,15,11,5
-28,16,11,5
-29,17,11,5
-30,14,12,3
-31,15,12,2
-32,16,12,3
-33,17,12,2
-34,22,13,3
-35,23,13,4
-36,24,13,5
-37,25,13,4
-38,22,14,2
-39,23,14,5
-40,24,14,5
-41,25,14,3
-42,39,15,3
-43,40,15,3
-44,41,15,3
-45,42,15,4
-46,39,16,4
-47,40,16,3
-48,41,16,2
-49,42,16,2
-50,39,17,3
-51,40,17,2
-52,41,17,1
-53,42,17,4
-54,39,18,2
-55,40,18,1
-56,41,18,4
-57,42,18,3
-58,39,19,1
-59,40,19,1
-60,41,19,2
-61,42,19,2
-62,39,20,2
-63,40,20,2
-64,41,20,2
-65,42,20,4
-66,43,21,3
-67,44,21,5
-68,45,21,2
-69,46,21,1
-70,43,22,2
-71,44,22,2
-72,45,22,3
-73,46,22,3
-74,43,23,5
-75,44,23,5
-76,45,23,2
-77,46,23,2
-78,43,24,4
-79,44,24,5
-80,45,24,3
-81,46,24,3
-82,43,25,2
-83,44,25,5
-84,45,25,2
-85,46,25,5
-86,47,26,1
-87,48,26,3
-88,49,26,1
-89,50,26,4
-90,47,27,2
-91,48,27,5
-92,49,27,4
-93,50,27,5
-94,47,28,3
-95,48,28,4
-96,49,28,3
-97,50,28,2
-98,47,29,1
-99,48,29,5
-100,49,29,1
-101,50,29,3
-102,47,30,5
-103,48,30,3
-104,49,30,5
-105,50,30,3
-106,47,31,2
-107,48,31,3
-108,49,31,4
-109,50,31,1
