--- src/data/testData/reviewsSimple.csv
@@ -1,51 +1,0 @@
-id,product_id,rating,recommend
-1,1,5,true
-2,1,4,false
-3,2,4,true
-4,2,4,true
-5,2,3,true
-6,2,5,false
-7,2,2,false
-8,4,4,true
-9,4,5,true
-10,4,2,false
-11,5,4,true
-12,5,3,true
-13,7,5,true
-14,7,1,false
-15,12,2,true
-16,12,1,false
-17,12,3,true
-18,12,2,true
-19,12,2,true
-20,12,2,true
-21,13,4,true
-22,13,1,false
-23,13,1,false
-24,13,1,false
-25,13,5,true
-26,14,3,true
-27,14,3,true
-28,14,5,true
-29,14,3,true
-30,14,1,false
-31,14,3,true
-32,15,1,false
-33,15,3,true
-34,15,4,true
-35,15,2,true
-36,16,5,true
-37,16,5,true
-38,16,5,true
-39,16,5,true
-40,16,1,false
-41,16,2,true
-42,16,3,true
-43,16,3,true
-44,16,4,true
-45,16,5,true
-46,16,4,true
-47,16,4,true
-48,16,1,false
-49,17,2,true
-50,17,2,true
